requirements.txt
pyjwt==2.8.0
cryptography>=42.0.0
python-dotenv==1.0.0
faker==20.1.0  # For generating decoy data
redis==5.0.1   # For production device cache
google-cloud-storage==2.13.0
google-cloud-secret-manager==2.16.4
pytest==7.4.3
pytest-cov==4.1.0
